agent/requirement.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
chromadb==1.5.4
click==8.3.1
colorlog==6.10.1
cryptography==46.0.5
dataclasses-json==0.6.7
defusedxml==0.7.1
dill==0.4.1
distlib==0.4.0
distro==1.9.0
docling==2.77.0
docling-core==2.69.0
docling-ibm-models==3.11.0
docling-parse==5.5.0
durationpy==0.10
et_xmlfile==2.0.0
Faker==40.8.0
filelock==3.25.0
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.49.0
google-genai==1.66.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.10
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.17
langchain-docling==2.0.0
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.14
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.3
oauthlib==3.3.1
ocrmac==1.0.1
omegaconf==2.3.0
onnxruntime==1.24.3
openai==2.26.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
polyfactory==3.3.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pylatexenc==2.10
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Vision==12.1
pypdf==6.7.5
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.1.1
python-docx==1.2.0
python-dotenv==1.2.2
python-pptx==1.0.2
pytz==2026.1.post1
PyYAML==6.0.3
rapidocr==3.7.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scipy==1.17.1
semchunk==3.2.5
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
virtualenv==21.1.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
